# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Fallstudie-Gruppe-3",
    "description": "API Dokumentation zur Anbindung von Frontends für das Kinoticketreservierungssystem."
  },
  "servers": [
    {
      "url": "https://fallstudie-gruppe-3.herokuapp.com/"
    },
    {
      "url": "http://fallstudie-gruppe-3.herokuapp.com/"
    }
  ],
...
  "paths": {
...
    "/vorstellungen": {
      "post": {
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "film": {
                    "example": "any"
                  },
                  "saal": {
                    "example": "any"
                  },
                  "startzeit": {
                    "example": "any"
                  },
                  "dauer": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "description": "Vorstellung anlegen.",
        "tags": [
          "Vorstellung"
        ]
      },
      "put": {
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vorstellungen"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Vorstellungen"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "saalnummer": {
                    "example": "any"
                  },
                  "startzeit": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "description": "Vorstellungen für einen Tag abrufen.",
        "tags": [
          "Vorstellung"
        ]
      }
    },
    "/vorstellungen/{saal}/{startzeit}": {
      "delete": {
        "parameters": [
          {
            "name": "saal",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startzeit",
            "in": "path",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                }
              }
            },
            "description": "Format: 2021-12-19T00:00"
          },
          {
            "name": "deleteShow",
            "in": "body",
            "description": "Zu entfernende Vorstellung.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/VorstellungLöschen"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "film": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "description": "Vorstellung löschen.",
        "tags": [
          "Vorstellung"
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "Vorstellungen": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "titel": {
              "type": "string",
              "example": "Django Unchained"
            },
            "startzeit": {
              "type": "string",
              "example": "2021-12-19T13:00"
            }
          }
        },
        "xml": {
          "name": "Vorstellungen"
        }
      },
...
      "VorstellungLöschen": {
        "type": "object",
        "properties": {
          "film": {
            "type": "string",
            "example": "Inception"
          }
        },
        "required": [
          "film"
        ],
        "xml": {
          "name": "VorstellungLöschen"
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}